server component related fixes
