+++ src/app/api/trpc/[trpc]/route.ts
@@ -22,9 +22,7 @@ const handler = (req: NextRequest) =>
     onError:
       env.NODE_ENV === "development"
         ? ({ path, error }) => {
-            console.error(
-              `❌ tRPC failed on ${path ?? "<no-path>"}: ${error.message}`,
-            );
+            console.error(`❌ tRPC failed on ${path ?? "<no-path>"}: ${error.message}`);
           }
         : undefined,
   });
