+++ src/server/api/trpc.ts
@@ -7,7 +7,7 @@
  * need to use are documented accordingly near the end.
  */
 
-import utils from "@/lib/auth/utils";
+import { unCachedValidateRequest } from "@/lib/auth";
 import { stripe } from "@/lib/stripe";
 import { db } from "@/server/db";
 import { initTRPC, TRPCError, type inferAsyncReturnType } from "@trpc/server";
@@ -27,7 +27,7 @@ import { ZodError } from "zod";
  * @see https://trpc.io/docs/server/context
  */
 export const createTRPCContext = async (opts: { headers: Headers }) => {
-  const { session, user } = await utils.validateRequest();
+  const { session, user } = await unCachedValidateRequest();
   return {
     session,
     user,
